survey now logs data into firebase
Upon finishing the survey, a score is calculated and put into the
database.

Right now the way the scores are calculated is that the tableview cells
are indexed from top to bottom starting at 0. Selecting a cell adds the
index number to the score. In the end the total score is pushed to
firebase.